Migrate to CoinGecko provider, modernize stack, fix CI, add mypy
Fixes the failing Daily Update workflow (CryptoCompare now key-gated → 401)
by migrating the data layer to CoinGecko behind a provider abstraction, and
brings the lib up to date.

Data layer:
- Add api.base.PriceProvider protocol + shared Coin dataclass and a
  get_price_provider() factory (PRICE_PROVIDER env, default coingecko).
- Add CoinGeckoClient (default): /coins/markets discovery, market_chart
  resampled to daily OHLCV (BTC/USD), rate limiting, optional Demo key.
- Keep CryptoCompare as the alternative provider; make fetcher/CLI/CI
  provider-agnostic. Scope the daily top-up to the top 300 coins.

Dependencies & tooling:
- Bump to current majors: pandas 3, numpy 2.2, scipy 1.18, plotly 6,
  tenacity 9, pytest 9, black 26, ruff 0.15, ipython 9; target-version py314.
- Commit poetry.lock for reproducible builds.
- Update GitHub Actions (checkout@v7, cache@v6).

CI:
- Refactor fetch-raw-data / calculate-total2 / pages into reusable
  (workflow_call) workflows; daily-update is now a thin orchestrator,
  removing ~1000 lines of duplicated bash.
- Add a mypy type-check gate and black --check to tests.yml; codebase is
  type-clean (0 errors).

Quality & docs:
- Fix code hygiene: pandas-3 set_index, narrowed exception handlers, cache
  write-error handling, surfaced symbol_replacements in TOTAL2 output,
  CLI global flags work before/after the subcommand.
- Fix doc inconsistencies (provider references, 4:00 AM schedule, Python
  3.14, round-trip window, filenames); rewrite DATA_SOURCES for CoinGecko;
  remove stale "free / no key" claims and historical phrasing.
- Add CoinGecko + provider tests and a CryptoCompare 401-path test.

+++ src/analysis/projections.py
@@ -194,14 +194,14 @@ def fit_log_trendlines(
             # Calculate intercept for upper line passing through the max2 point (or average if multiple)
             if major_peaks:
                 # Use the major peak(s) to set the upper intercept
-                major_peak_x = np.mean(
-                    [(get_regression_date(p) - reference_date).days for p in major_peaks]
+                major_peak_x = float(
+                    np.mean([(get_regression_date(p) - reference_date).days for p in major_peaks])
                 )
                 major_peak_y = np.mean([np.log10(p.price) for p in major_peaks])
             else:
                 # Fallback to highest peak
                 highest_peak = max(peaks, key=lambda p: p.price)
-                major_peak_x = (get_regression_date(highest_peak) - reference_date).days
+                major_peak_x = float((get_regression_date(highest_peak) - reference_date).days)
                 major_peak_y = np.log10(highest_peak.price)
             upper_intercept = major_peak_y - slope * major_peak_x
             return (
@@ -217,14 +217,14 @@ def fit_log_trendlines(
             slope = upper_fit[0]
             # Calculate intercept for lower line passing through the min1 point (or average if multiple)
             if major_troughs:
-                major_trough_x = np.mean(
-                    [(get_regression_date(p) - reference_date).days for p in major_troughs]
+                major_trough_x = float(
+                    np.mean([(get_regression_date(p) - reference_date).days for p in major_troughs])
                 )
                 major_trough_y = np.mean([np.log10(p.price) for p in major_troughs])
             else:
                 # Fallback to lowest trough
                 lowest_trough = min(troughs, key=lambda p: p.price)
-                major_trough_x = (get_regression_date(lowest_trough) - reference_date).days
+                major_trough_x = float((get_regression_date(lowest_trough) - reference_date).days)
                 major_trough_y = np.log10(lowest_trough.price)
             lower_intercept = major_trough_y - slope * major_trough_x
             return (
